# optimize CombineVHostRoutes