Moves the expensive sorting part of the merge after all appends are complete. 
Goes from m*n to m+n for merging. 

<table class='benchstat '>
<tr class='configs'><th><th>master (7d9e726d4c)<th>548e4acba0<th>06047384a7</th>
<tbody>
<tr><th><th colspan='3' class='metric'>time/op
<tr>
    <td>RouteGeneration/gateways-12<td>38.0ms ± 1%<td>38.4ms ± 2%<td>40.1ms ± 1%
<tr>
    <td><b>RouteGeneration/gateways-shared-12</b><td><b style="color: red">364ms ± 7%</b><td><b>173ms ± 2%</b><td><b>177ms ± 1%</b>
<tr>
    <td>RouteGeneration/empty-12<td>676µs ± 2%<td>680µs ± 2%<td>679µs ± 1%
<tr>
    <td>RouteGeneration/telemetry-12<td>686µs ± 1%<td>691µs ± 1%<td>695µs ± 2%
<tr>
    <td>RouteGeneration/virtualservice-12<td>1.57ms ± 2%<td>1.55ms ± 1%<td>1.55ms ± 1%
</tbody>
</table>

### 3 -> 12 hosts

**master vs new (with 12)**
```
name                                old time/op        new time/op        delta
RouteGeneration/gateways-shared-12         1.93s ±13%         0.60s ± 3%  -68.88%  (p=0.008 n=5+5)
```

**master with 3 vs 12**
```
RouteGeneration/gateways-shared-12         364ms ± 7%        1930ms ±13%  +429.87%  (p=0.008 n=5+5)
```

**new with 3 vs 12 **
```
RouteGeneration/gateways-shared-12         173ms ± 2%         601ms ± 3%  +247.63%  (p=0.008 n=5+5)
```

### 1000 -> 2000 Services
**master with 1000 vs 2000**
```
RouteGeneration/gateways-shared-12         372ms ± 2%        1240ms ±17%  +233.60%  (p=0.008 n=5+5)
```

**new with 1000 vs 2000**
```
RouteGeneration/gateways-shared-12         178ms ± 3%         360ms ± 2%  +102.25%  (p=0.008 n=5+5)
```

[ ] Configuration Infrastructure
[ ] Docs
[ ] Installation
[] Networking
[X] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[ ] User Experience
[ ] Developer Infrastructure