@istio-testing 2020-07-10
@stevenctl: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-security-k8s-tests_istio | 548e4acba0ad4e8391c4b7787621956f0b1822aa | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25378/integ-security-k8s-tests_istio/15975) | `/test integ-security-k8s-tests_istio`
integ-mixer-k8s-tests_istio | 548e4acba0ad4e8391c4b7787621956f0b1822aa | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25378/integ-mixer-k8s-tests_istio/14178) | `/test integ-mixer-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->