@stevenctl 2020-07-10
/test integ-distroless-k8s-tests_istio